requirements.txt
dico-api
sanic
PyNaCl